Bluetooth: Factor out bt_data into its own header
Move struct bt_data, related macros (BT_DATA, BT_DATA_BYTES,
BT_DATA_SERIALIZED_SIZE) and helper function declarations
(bt_data_get_len, bt_data_serialize) from bluetooth.h into a new
include/zephyr/bluetooth/data.h header.

This resolves the issue where bluetooth.h includes classic.h before
struct bt_data is defined. Now classic.h can directly include data.h
instead of relying on a forward declaration. bluetooth.h still includes
data.h for backwards compatibility.
